+++ topic/topicoptions/topicoptions_describe.go
@@ -7,6 +7,4 @@ import "github.com/ydb-platform/ydb-go-sdk/v3/internal/grpcwrapper/rawtopic"
 // # Experimental
 //
 // Notice: This API is EXPERIMENTAL and may be changed or removed in a later release.
-type DescribeOption interface {
-	ApplyDescribeOption(req *rawtopic.DescribeTopicRequest)
-}
+type DescribeOption func(req *rawtopic.DescribeTopicRequest)